Do not install packages that already exist - ansible 2.8
There is a difference in behavior of the `package` (`dnf`) module
between ansible 2.8 and 2.9.  If you do this on 2.8:
```yaml
- name: install
  package:
    name: somepackage
    state: present
```
It will upgrade the package if it already exists and there is a newer
version.  Ansible 2.9 will not upgrade the package.

This causes problems for crypto-policies because we do not want to
alter the state of the system if we do not need to.  The solution is
to use `package_facts` to get the list of packages on the target
system, and only install the package from `__crypto_policies_packages`
if it does not exist.  It is the responsibility of the user to update
the packages on the managed nodes if the user wants to use the latest
packages.

Note that `dnf -y install somepackage` on EL8 will update the package
if it exists, so the `dnf` module in Ansible 2.9 has a different
behavior than the `dnf` command.