docs: remove unused Dependencies section in README
remove unused Dependencies section in README - it is confusing
to users who look at the README, see the `Dependencies` section
that says `None`, and think that there are no dependencies for
the role, when the actual dependencies are listed in the
`Requirements` section which is not even near the `Dependencies`
section.  So consistently use the `Requirements` section and
get rid of the `Dependencies` section.

Fix some markdownlint and ansible-lint issues

Signed-off-by: Rich Megginson <rmeggins@redhat.com>